+++ src/DI/Definition/Resolver/ValueDefinitionResolver.php
@@ -27,15 +27,13 @@ class ValueDefinitionResolver implements DefinitionResolver
      *
      * {@inheritdoc}
      */
-    public function resolve(Definition $definition)
+    public function resolve(Definition $definition, array $parameters = array())
     {
         if (! $definition instanceof ValueDefinition) {
-            throw new \InvalidArgumentException(
-                sprintf(
-                    'This definition resolver is only compatible with ValueDefinition objects, %s given',
-                    get_class($definition)
-                )
-            );
+            throw new \InvalidArgumentException(sprintf(
+                'This definition resolver is only compatible with ValueDefinition objects, %s given',
+                get_class($definition)
+            ));
         }
 
         return $definition->getValue();